Added plotting lib for basic plots
