+++ pyproject.toml
@@ -18,7 +18,7 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
     "Operating System :: OS Independent",
 ]
-version = "0.0.1.1"
+version = "0.0.1.2"
 dependencies = [
     "pandas",
     "binance-connector",
@@ -26,5 +26,7 @@ dependencies = [
     "pyyaml",
     "yfinance",
     "alpaca-trade-api",
+    "plotly",
+    "kaleido",
     "ta"
 ]
