+++ tests/test_technicals.py
@@ -8,7 +8,6 @@ import numpy as np
 b = BinanceCandleData()
 b._set_params('BTCUSDT', b.HOUR_1, 1000, None, None)
 b.df = pd.read_pickle('tests/data/btc_test_data')
-b._set_candle_gap()
 t = Technicals(b.df, peak_spacing=20)
 
 def test_price_dips():
